move markdown parser to common. add unit test
use `markdown-it-py` instead of `markdown` dependency for test comparison since it's already in Pipfile.lock
