+++ Pipfile
@@ -42,6 +42,7 @@ jupyterlab = "*"
 jupyterlab-vim = "*"
 keras_applications = "*"
 lru-dict = "*"
+markdown-it-py = "*"
 matplotlib = "*"
 mock = "*"
 mpld3 = "*"
@@ -95,7 +96,6 @@ hexdump = "*"
 Jinja2 = "*"
 json-rpc = "*"
 libusb1 = "*"
-markdown = "*"
 nose = "*"
 numpy = "*"
 onnx = "*"
