+++ selfdrive/updated.py
@@ -35,6 +35,7 @@ from pathlib import Path
 from typing import List, Tuple, Optional
 
 from common.basedir import BASEDIR
+from common.markdown import parse_markdown
 from common.params import Params
 from selfdrive.hardware import EON, TICI, HARDWARE
 from selfdrive.swaglog import cloudlog
@@ -95,44 +96,6 @@ def set_consistent_flag(consistent: bool) -> None:
   elif not consistent:
     consistent_file.unlink(missing_ok=True)
   os.sync()
-
-
-def parse_markdown(text: str, tab_length: int = 2) -> str:
-  lines = text.split("\n")
-  output: List[str] = []
-  list_level = 0
-
-  def end_outstanding_lists(level: int, end_level: int) -> int:
-    while level > end_level:
-      level -= 1
-      output.append("</ul>")
-      if level > 0:
-        output.append("</li>")
-    return end_level
-
-  for i, line in enumerate(lines):
-    if i+1 < len(lines) and lines[i+1].startswith("==="): # heading
-      output.append(f"<h1>{line}</h1>")
-    elif line.startswith("==="):
-      pass
-    elif line.lstrip().startswith("* "): # list
-      line_level = 1 + line.count(" " * tab_length, 0, line.index("*"))
-      if list_level >= line_level:
-        list_level = end_outstanding_lists(list_level, line_level)
-      else:
-        list_level += 1
-        if list_level == 1:
-          output.append("<ul>")
-        else:
-          output[-1] = output[-1].replace("</li>", "<ul>")
-      output.append(f"<li>{line.replace('*', '', 1).lstrip()}</li>")
-    else: # catchall
-      list_level = end_outstanding_lists(list_level, 0)
-      if len(line) > 0:
-        output.append(line)
-
-  end_outstanding_lists(list_level, 0)
-  return '\n'.join(output)
 
 
 def set_params(new_version: bool, failed_count: int, exception: Optional[str]) -> None:
